Refactor main.py and add build workflow
Refactor `main.py` by moving functions to separate files based on their tasks and add a GitHub Actions workflow to generate binary files for all platforms.

* **Refactor Functions:**
  - Move `figlet_header()` to `header.py`.
  - Move `list_available_profiles()` and `select_profile()` to `profile.py`.
  - Move `loading_animation()` to `animation.py`.
  - Move `check_service_in_region()` and `list_active_regions()` to `region_check.py`.
  - Move `get_topic_attributes()`, `list_subscriptions()`, `send_message()`, `brute_force_regions()`, and `check_sns_misconfigurations()` to `sns_operations.py`.
  - Move `save_state()` and `load_state()` to `state.py`.
  - Update `main.py` to import functions from the new files and retain only the `main()` function.

* **Add GitHub Actions Workflow:**
  - Add `build.yml` in `.github/workflows/` to build the project for Windows, macOS, and Linux platforms.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/harekrishnarai/Snare?shareId=XXXX-XXXX-XXXX-XXXX).
